refactor: split docs extractor driver helpers
